pvbatch mode with MPI for grid2paraview.py
Changed grid2paraview.py to operate with MPI
when invoked with pvbatch -sym. Command line
syntax below.

mpiexec -np <num procs> pvbatch -sym grid2paraview \
  <same options as original pvpython version>

Must use the mpiexec and pvbatch that are in the
same ParaView/bin directory as pvpython (or an mpiexec
that works with your ParaView install). The original
pvpython mode operates the same as before.